refactor(main): put one Team API call behind one client
Making a request meant reading eight of the manager's methods: `request`, `#fetch`,
`#requestProtocol`, `#ensureCompatibility`, two negotiators, `#verifyIdentity` and
`#probeRemoteDesktop`, threaded through a private field and reachable only through the whole
manager. `RemoteServerClient` is those eight, with the three things they need named as interfaces --
a server lookup, a place to record what happened, and a two-method transport.

That last one is the point of the file. `RemoteHostRequestTransport` declares `request` and
`requestResponse` and nothing else, so the live event channel and the control-plane client that
`TeamWebRtcClientTransport` also carries are structurally out of reach here. The direction
established in the previous commit now holds in code: a request reports a failure to the registry,
and this file does not know a WebSocket exists.

The header says what must not be tidied. The two arms of every method are two released wire
protocols -- V1 or V3 over HTTPS by negotiation, V2 JSON over WebRTC -- that happen to carry the
same payload today, and merging them is the one edit here that breaks a build already installed.
`packages/contracts/src/team-api-routes.ts` now points at this file as the client half of that
surface, since it is no longer the manager.

`#compatibilityRequests` moved inside the client, where it belongs: it is in-flight deduplication
the UI never sees, and it was the reason `#clearServerConnectionState` mixed a private concern in
with the four the UI does see.

Behaviour is unchanged with one seam made explicit: `refreshWebRtcCompatibility` records the
compatibility and the manager emits `changed`, where before both happened inside one method.

Watch it fail: dropping `reportError` from the request path reddens "closes the event data plane
after a malformed HTTP payload" (`expected "vi.fn()" to be called with arguments: [ 1000, 'Client
stopped' ]`) -- the request-to-registry-to-socket hook survived the split; letting a recorded
protocol issue renegotiate reddens "blocks a host range with no shared protocol" (`expected
"vi.fn()" to be called 1 times, but got 2 times`).

Manager: 1736 -> 1468 lines. Step 8 of 14.
